chore: refactoring and documentation
